+++ package.json
@@ -61,6 +61,7 @@
     "rollup-plugin-sourcemaps": "^0.4.2",
     "rollup-plugin-uglify": "^3.0.0",
     "slate-collapse-on-escape": "^0.6.0",
+    "slate-dev-benchmark": "*",
     "slate-soft-break": "^0.6.0",
     "source-map-loader": "^0.2.3",
     "source-map-support": "^0.4.0",
@@ -80,7 +81,7 @@
   "scripts": {
     "benchmark": "mkdir -p ./tmp && node ./node_modules/.bin/_matcha --reporter ./support/benchmark/reporter ./tmp/benchmark/*/index.js > ./tmp/benchmark-comparison.json && cross-env BABEL_ENV=test babel-node ./support/benchmark/compare",
     "benchmark:save": " yarn benchmark:prepare && node ./node_modules/.bin/_matcha --reporter ./support/benchmark/reporter ./tmp/benchmark/*/index.js > ./tmp/benchmark-baseline.json",
-    "benchmark:prepare": "mkdir -p ./tmp && cross-env BABEL_ENV=benchmark babel benchmark --out-dir tmp/benchmark/ && cp support/benchmark/ugly-fix/* node_modules/matcha/lib/matcha/timer/",
+    "benchmark:prepare": "mkdir -p ./tmp && cross-env BABEL_ENV=test babel benchmark --out-dir tmp/benchmark/ && cp support/benchmark/ugly-fix/* node_modules/matcha/lib/matcha/timer/",
     "bootstrap": "lerna bootstrap && yarn build",
     "build": "rollup --config ./support/rollup/config.js",
     "build:production": "cross-env NODE_ENV=production rollup --config ./support/rollup/config.js && cross-env NODE_ENV=production webpack --config support/webpack/config.js",
